pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "oto-cli"
dynamic = ["version"]
description = "CLI toolkit for AI agents — covers the long tail of SaaS APIs that don't have a CLI"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Alexis Laporte" }
]
keywords = ["cli", "ai-agents", "automation", "api", "llm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]
dependencies = [
    "typer>=0.9.0",
    "requests>=2.28.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
google = [
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "google-api-python-client>=2.0.0",
    "gkeepapi>=0.16.0",
]
browser = [
    "o-browser",
]
anthropic = [
    "anthropic>=0.40.0",
]
stock = [
    "pyarrow>=10.0.0",
    "pandas>=1.5.0",
]
all = [
    "oto-cli[google,browser,anthropic,stock]",
]

[project.urls]
Repository = "https://github.com/AlexisLaporte/oto"

[project.scripts]
oto = "oto.cli:main"

[tool.hatch.version]
path = "oto/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["oto"]
